Jstree Updated
Added bootstrap css to the jstree. Modified the tree so it no longer utilizes .on(select_node) and the user now has to utilize the context menu to 'query' nodes for further information / separate pages. The items Create New, Rename, and Delete were added to the context menu and can be called functions later in the <script>.