fixed JSTree
The State bug was actually an issue with the idea of the JSON that makes up the JStree. Essentially, when the tree is made, it creates its own JSON file only made up of its premade distinct values. I was incorrectly calling on these premade values so the system kept getting confused as to what asset or component id was actually right (all were either 4 or 6) which loads the two GEICO components respectively. The fix was to call node.original.id as the ‘original’ field is the JSON I create for the tree.