Redirect after submission fixed
The redirection to the previous unit page after creating/editing an entry now results in page redirection. Previous iteration did not work because, since I'm using an ajax request to send the POST request, I need to return a JSON response of success and then load the new page in by changing the window.location in the ajax's success function. Ajax does not allow the page to be changed even if the proper view function is called and a GET request is made.