Implemented Asset Page Generation w/ JStree
JStree can now be used to navigate to an asset page based on the specific asset queried. Because Assets are foreign keys, the implementation also calls for the now OneToOne field component to ensure only components from that asset instance are called.

Also added JSDataTable sorting for the closed? checkboxes.